trialtwin-lab/requirements.txt
numpy>=1.24,<2.0
pandas>=2.0,<3.0
lifelines>=0.27,<1.0
scikit-survival>=0.22,<1.0
scikit-learn>=1.3,<2.0
scipy>=1.11,<2.0
streamlit>=1.28,<2.0
plotly>=5.18,<6.0
matplotlib>=3.8,<4.0
pyyaml>=6.0,<7.0
click>=8.1,<9.0
pytest>=7.4,<9.0
